--- a/.travis.yml
+++ b/.travis.yml
@@ -9,9 +9,11 @@ cache:
   directories:
     - $HOME/zsh
     - $HOME/.zsh-builds
+    - $HOME/.antigen
 
 env:
   global:
+    - ANTIGEN=$TRAVIS_BUILD_DIR
     - ZSH_REMOTE_URL=https://github.com/zsh-users/zsh.git
     - BUILDS_PATH=$HOME/.zsh-builds
     - ZSH_SOURCE=$HOME/zsh
@@ -22,6 +24,13 @@ env:
     - ZSH_VERSION=zsh-5.0.0
     - ZSH_VERSION=zsh-4.3.17
     - ZSH_VERSION=zsh-4.3.5
+
+install:
+  # Install python requirements.
+  - "pip install cram==0.6"
+
+  # Install zsh versions from source.
+  - "./tests/build-zsh.sh $ZSH_VERSION $BUILDS_PATH $ZSH_SOURCE"
 
 before_script:
   # Show the git version being used to test.
@@ -34,12 +43,9 @@ before_script:
   # Show zsh version being deployed.
   - "$BUILDS_PATH/$ZSH_VERSION/zsh --version"
 
-install:
-  # Install python requirements.
-  - "pip install cram==0.6"
-
-  # Install zsh versions from source.
-  - "./tests/build-zsh.sh $ZSH_VERSION $BUILDS_PATH $ZSH_SOURCE"
-
 # Run the tests.
-script: "make tests PYENV= SHELL=$BUILDS_PATH/$ZSH_VERSION/zsh"
+script:
+  - "make tests PYENV= SHELL=$BUILDS_PATH/$ZSH_VERSION/zsh"
+  - "make stats _ANTIGEN_CACHE_ENABLED=false _ANTIGEN_INIT_ENABLED=false SHELL=$BUILDS_PATH/$ZSH_VERSION/zsh"
+  - "make stats _ANTIGEN_CACHE_ENABLED=true _ANTIGEN_INIT_ENABLED=false SHELL=$BUILDS_PATH/$ZSH_VERSION/zsh"
+  - "make stats _ANTIGEN_CACHE_ENABLED=true _ANTIGEN_INIT_ENABLED=true SHELL=$BUILDS_PATH/$ZSH_VERSION/zsh"