--- a/antigen.zsh
+++ b/antigen.zsh
@@ -7,11 +7,12 @@
 # Each line in this string has the following entries separated by a space
 # character.
 # <repo-url>, <plugin-location>, <bundle-type>, <has-local-clone>
-# FIXME: Is not kept local by zsh!
 local _ANTIGEN_BUNDLE_RECORD=""
 local _ANTIGEN_INSTALL_DIR="$(cd "$(dirname "$0")" && pwd)"
+local _ANTIGEN_CACHE_ENABLED=${_ANTIGEN_CACHE_ENABLED:-true}
+local _ZCACHE_EXTENSION_LOADED=false
 
-# Do not load anything if anything is no available.
+# Do not load anything if git is no available.
 if ! which git &> /dev/null; then
     echo 'Antigen: Please install git to use Antigen.' >&2
     return 1
@@ -21,12 +22,63 @@ fi
 typeset -a __deferred_compdefs
 compdef () { __deferred_compdefs=($__deferred_compdefs "$*") }
 
+-antigen-parse-bundle () {
+  # Bundle spec arguments' default values.
+  local url="$ANTIGEN_DEFAULT_REPO_URL"
+  local loc=/
+  local branch=
+  local no_local_clone=false
+  local btype=plugin
+
+  # Parse the given arguments. (Will overwrite the above values).
+  eval "$(-antigen-parse-args \
+          'url?, loc? ; branch:?, no-local-clone?, btype:?' \
+          "$@")"
+
+  # Check if url is just the plugin name. Super short syntax.
+  if [[ "$url" != */* ]]; then
+      loc="plugins/$url"
+      url="$ANTIGEN_DEFAULT_REPO_URL"
+  fi
+
+  # Resolve the url.
+  url="$(-antigen-resolve-bundle-url "$url")"
+
+  # Add the branch information to the url.
+  if [[ ! -z $branch ]]; then
+      url="$url|$branch"
+  fi
+
+  # The `make_local_clone` variable better represents whether there should be
+  # a local clone made. For cloning to be avoided, firstly, the `$url` should
+  # be an absolute local path and `$branch` should be empty. In addition to
+  # these two conditions, either the `--no-local-clone` option should be
+  # given, or `$url` should not a git repo.
+  local make_local_clone=true
+  if [[ $url == /* && -z $branch &&
+          ( $no_local_clone == true || ! -d $url/.git ) ]]; then
+      make_local_clone=false
+  fi
+
+  # Add the theme extension to `loc`, if this is a theme.
+  if [[ $btype == theme && $loc != *.zsh-theme ]]; then
+      loc="$loc.zsh-theme"
+  fi
+
+  # Bundle spec arguments' default values.
+  echo "local url=\""$url\""
+        local loc=\""$loc\""
+        local branch=\""$branch\""
+        local make_local_clone="$make_local_clone"
+        local btype=\""$btype\""
+        "
+}
+
 # Syntaxes
 #   antigen-bundle <url> [<loc>=/]
 # Keyword only arguments:
 #   branch - The branch of the repo to use for this bundle.
 antigen-bundle () {
-
     # Bundle spec arguments' default values.
     local url="$ANTIGEN_DEFAULT_REPO_URL"
     local loc=/
@@ -34,40 +86,7 @@ antigen-bundle () {
     local no_local_clone=false
     local btype=plugin
 
-    # Parse the given arguments. (Will overwrite the above values).
-    eval "$(-antigen-parse-args \
-            'url?, loc? ; branch:?, no-local-clone?, btype:?' \
-            "$@")"
-
-    # Check if url is just the plugin name. Super short syntax.
-    if [[ "$url" != */* ]]; then
-        loc="plugins/$url"
-        url="$ANTIGEN_DEFAULT_REPO_URL"
-    fi
-
-    # Resolve the url.
-    url="$(-antigen-resolve-bundle-url "$url")"
-
-    # Add the branch information to the url.
-    if [[ ! -z $branch ]]; then
-        url="$url|$branch"
-    fi
-
-    # The `make_local_clone` variable better represents whether there should be
-    # a local clone made. For cloning to be avoided, firstly, the `$url` should
-    # be an absolute local path and `$branch` should be empty. In addition to
-    # these two conditions, either the `--no-local-clone` option should be
-    # given, or `$url` should not a git repo.
-    local make_local_clone=true
-    if [[ $url == /* && -z $branch &&
-            ( $no_local_clone == true || ! -d $url/.git ) ]]; then
-        make_local_clone=false
-    fi
-
-    # Add the theme extension to `loc`, if this is a theme.
-    if [[ $btype == theme && $loc != *.zsh-theme ]]; then
-        loc="$loc.zsh-theme"
-    fi
+    eval "$(-antigen-parse-bundle "$@")"
 
     # Add it to the record.
     _ANTIGEN_BUNDLE_RECORD="$_ANTIGEN_BUNDLE_RECORD\n$url $loc $btype"
@@ -107,9 +126,7 @@ antigen-bundles () {
     # Bulk add many bundles at one go. Empty lines and lines starting with a `#`
     # are ignored. Everything else is given to `antigen-bundle` as is, no
     # quoting rules applied.
-
     local line
-
     grep '^[[:space:]]*[^[:space:]#]' | while read line; do
         # Using `eval` so that we can use the shell-style quoting in each line
         # piped to `antigen-bundles`.
@@ -259,66 +276,71 @@ antigen-revert () {
 
 }
 
+-antigen-load-list () {
+  local url="$1"
+  local loc="$2"
+  local make_local_clone="$3"
+  local sources=''
+
+  # The full location where the plugin is located.
+  local location
+  if $make_local_clone; then
+      location="$(-antigen-get-clone-dir "$url")/"
+  else
+      location="$url/"
+  fi
+
+  [[ $loc != "/" ]] && location="$location$loc"
+
+  if [[ -f "$location" ]]; then
+      sources="$location"
+  else
+
+      # Source the plugin script.
+      # FIXME: I don't know. Looks very very ugly. Needs a better
+      # implementation once tests are ready.
+      local script_loc="$(ls "$location" | grep '\.plugin\.zsh$' | head -n1)"
+
+      if [[ -f $location/$script_loc ]]; then
+          # If we have a `*.plugin.zsh`, source it.
+          sources="$location/$script_loc"
+
+      elif [[ -f $location/init.zsh ]]; then
+          # Otherwise source it.
+          sources="$location/init.zsh"
+
+      elif ls "$location" | grep -l '\.zsh$' &> /dev/null; then
+          # If there is no `*.plugin.zsh` file, source *all* the `*.zsh`
+          # files.
+
+          for script ($location/*.zsh(N)) {
+            sources="$sources\n$script"
+          }
+
+      elif ls "$location" | grep -l '\.sh$' &> /dev/null; then
+          # If there are no `*.zsh` files either, we look for and source any
+          # `*.sh` files instead.
+          for script ($location/*.sh(N)) {
+            sources="$sources\n$script"
+          }
+      fi
+  fi
+
+  echo "$sources"
+}
+
 -antigen-load () {
-
-    local url="$1"
-    local loc="$2"
-    local make_local_clone="$3"
-
-    # The full location where the plugin is located.
-    local location
-    if $make_local_clone; then
-        location="$(-antigen-get-clone-dir "$url")/"
+  local src
+  for src in $(-antigen-load-list "$1" "$2" "$3"); do
+    # Add to $fpath, for completion(s), if not in there already
+    if [[ -d "$src" ]]; then
+      if (( ! ${fpath[(I)$location]} )); then
+         fpath=($location $fpath)
+      fi
     else
-        location="$url/"
+      source "$src"
     fi
-
-    [[ $loc != "/" ]] && location="$location$loc"
-
-    if [[ -f "$location" ]]; then
-        source "$location"
-
-    else
-
-        # Source the plugin script.
-        # FIXME: I don't know. Looks very very ugly. Needs a better
-        # implementation once tests are ready.
-        local script_loc="$(ls "$location" | grep '\.plugin\.zsh$' | head -n1)"
-
-        if [[ -f $location/$script_loc ]]; then
-            # If we have a `*.plugin.zsh`, source it.
-            source "$location/$script_loc"
-
-        elif [[ -f $location/init.zsh ]]; then
-            # If we have a `init.zsh`
-            if (( $+functions[pmodload] )); then
-                # If pmodload is defined pmodload the module. Remove `modules/`
-                # from loc to find module name.
-                pmodload "${loc#modules/}"
-            else
-                # Otherwise source it.
-                source "$location/init.zsh"
-            fi
-
-        elif ls "$location" | grep -l '\.zsh$' &> /dev/null; then
-            # If there is no `*.plugin.zsh` file, source *all* the `*.zsh`
-            # files.
-            for script ($location/*.zsh(N)) { source "$script" }
-
-        elif ls "$location" | grep -l '\.sh$' &> /dev/null; then
-            # If there are no `*.zsh` files either, we look for and source any
-            # `*.sh` files instead.
-            for script ($location/*.sh(N)) { source "$script" }
-
-        fi
-
-        # Add to $fpath, for completion(s), if not in there already
-        if (( ! ${fpath[(I)$location]} )); then
-            fpath=($location $fpath)
-        fi
-
-    fi
-
+  done
 }
 
 # Update (with `git pull`) antigen itself.
@@ -434,7 +456,6 @@ antigen-prezto-lib () {
 }
 
 antigen-theme () {
-
     if [[ "$1" != */* && "$1" != --* ]]; then
         # The first argument is just a name of the plugin, to be picked up from
         # the default repo.
@@ -565,9 +586,22 @@ documentation, visit the project's page at 'http://antigen.sharats.me'.
 EOF
 }
 
+# Load zcache extension if not already loaded
+-antigen-load-extension () {
+    if ! $_ZCACHE_EXTENSION_LOADED; then
+        #_ZCACHE_EXTENSION_LOADED is set to true in zcache.zsh
+        source "$_ANTIGEN_INSTALL_DIR/ext/zcache.zsh"
+        zcache-start
+    fi
+}
+
 # A syntax sugar to avoid the `-` when calling antigen commands. With this
 # function, you can write `antigen-bundle` as `antigen bundle` and so on.
 antigen () {
+    if $_ANTIGEN_CACHE_ENABLED; then
+        -antigen-load-extension
+    fi
+
     local cmd="$1"
     if [[ -z "$cmd" ]]; then
         echo 'Antigen: Please give a command to run.' >&2
@@ -745,8 +779,8 @@ antigen () {
     -set-default ADOTDIR $HOME/.antigen
 
     # Setup antigen's own completion.
-    autoload -U compinit
-    compinit -C
+    autoload -Uz compinit
+    #compinit -C
     compdef _antigen antigen
 
     # Remove private functions.
@@ -770,8 +804,11 @@ _antigen () {
     'snapshot:Create a snapshot of all the active clones'
     'restore:Restore the bundles state as specified in the snapshot'
     'selfupdate:Update antigen itself'
-    'help:Print help message'
-  )
+  );
+
+  $_ANTIGEN_CACHE_ENABLED && _1st_arguments+=('cache-reset:Clears bundle cache')
+
+  _1st_arguments+=('help:Show this message')
 
   __bundle() {
     _arguments \