--- a/tests/.zshenv
+++ b/tests/.zshenv
@@ -4,6 +4,9 @@
 
 export ADOTDIR="$PWD/dot-antigen"
 
+export _ANTIGEN_CACHE_PATH="$TESTDIR/.cache"
+export _ANTIGEN_CACHE_ENABLED=false
+
 test -f "$TESTDIR/.zcompdump" && rm "$TESTDIR/.zcompdump"
 
 source "$TESTDIR/../antigen.zsh"
@@ -34,6 +37,7 @@ mkdir "$PLUGIN_DIR2"
 alias pg2='git --git-dir "$PLUGIN_DIR2/.git" --work-tree "$PLUGIN_DIR2"'
 
 echo 'alias hehe2="echo hehe2"' > "$PLUGIN_DIR2"/init.zsh
+echo -E 'alias prompt="\e]$ >\a\n"' >> "$PLUGIN_DIR2"/init.zsh
 echo 'alias unsourced-alias="echo unsourced-alias"' > "$PLUGIN_DIR2"/aliases.zsh
 
 {