Basic cache implementation following the description given [here](https://github.com/zsh-users/antigen/issues/119).

Very *raw* implementation. All feedback truly appreciated!

To enable/disable set _ANTIGEN_CACHE_ENABLED environment variable apropiately: true or false.

To configure cache path use _ANTIGEN_CACHE_DIR environment variable. By default it's _ANTIGEN_INSTALL_DIR/.cache.

Sample load times on a Intel Core i7 @ 2.40Ghz x 8 64 bit with SSD, antigen configured with [11 extensions](https://github.com/desyncr/zshrc/blob/master/.antigenrc):

    # First time (cache enabled, generates cache)
    $ time zsh
    zsh  0.45s user 0.67s system 119% cpu 0.932 total

    # Second time (cache enable, load from cache)
    $ time zsh
    zsh  0.35s user 0.31s system 113% cpu 0.574 total

    # Cache disabled
    $ export _ANTIGEN_CACHE_ENABLED=false
    $ time zsh
    zsh  0.41s user 0.69s system 119% cpu 0.925 total

## TODO

- [x] ~~Clean cache when bundle is updated/modified~~ Unfeasible. Added commands 'cache-rebuild' and 'cache-clean' for that.
- [x] ~~Intercept expensive calls/functions rather than modify them to add calls to cache-start/cache-done~~ See https://github.com/zsh-users/antigen/pull/133
- [x] Support prezto plugins (See [this](https://github.com/zsh-users/antigen/pull/129/files#diff-af5e58dd8e0859139fb7af9bd8034d85R45))
- [x] Fix all code style issues (indentation, naming, env variables...)
- [x] Add command to rebuild cache, enable/disable cache
- [x] Fix clear-cache messages (extra empty line etc)
- [x] Manage 'antigen bundle' command
- [x] Test on Fedora, CentOS, Debian, Ubuntu, Archlinux, Mac OS X Yosemite. El Capitan
- [x] Remove confirmation prompt for cache-reset command
- [x] Test changing themes, update, revert, snapshot, restore