@desyncr 2015-06-11
```
$ time zsh
zsh  0.30s user 0.64s system 121% cpu 0.773 total

$ export _ANTIGEN_CACHE_ENABLED=true
$ time zsh
zsh  0.30s user 0.71s system 124% cpu 0.806 total

$ time zsh
zsh  0.12s user 0.20s system 115% cpu 0.276 total
```

![](http://i0.kym-cdn.com/photos/images/original/000/119/924/sweetjesus.jpg)