@desyncr 2015-06-13
```
$ export _ANTIGEN_CACHE_ENABLED=false
$ time zsh
zsh  0.20s user 0.63s system 127% cpu 0.655 total

$ export _ANTIGEN_CACHE_ENABLED=true
$ time zsh
zsh  0.32s user 0.87s system 126% cpu 0.941 total

$ time zsh
zsh  0.10s user 0.12s system 111% cpu 0.202 total
```