@forivall 2015-06-13
I've found two issues with this:
1. using `sed -i` sometimes changes the permissions on paths with special permissions (ACLS). The following diff will fix it.
```
diff --git a/ext/zcache.zsh b/ext/zcache.zsh
index 7ecfe1f..f33e762 100644
--- a/ext/zcache.zsh
+++ b/ext/zcache.zsh
@@ -158,9 +158,8 @@ function -zcache-done () {
 
     # TODO add option
     # if $_ANTIGEN_CACHE_MINIFY; then
-        sed -i '/^#.*/d' $_zcache_payload_path
-        sed -i '/^$/d' $_zcache_payload_path
-        sed -i '/./!d' $_zcache_payload_path
+        sed -i -e '/^#.*/d' -e '/^$/d' -e '/./!d' $_zcache_payload_path
+        chmod 644 $_zcache_payload_path
     # fi
 
     -dots-stop-capture $_zcache_meta_path
```

2. Plugins that `source` via relative paths are broken. I'd fix it by caching the file referenced with the source if it's a static source, and if there are dynamic `source`s, create a function that will intercept the `source` and change to the plugin's original directory before running `source`.