@desyncr 2015-06-14
@forivall Regarding 1. Can you specify the scenarios where this behavior is shown, as so to add tests.

Regarding 2. This behavior is shown in -antigen-load method. Not the cache implementation. Raise an issue there and changes would be backported.